// frontend/src/api/generated/models/paymentRequestPayload.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface PaymentRequestPayload {
  request_type: string;
  source_type: string;
  source_id: string;
  amount: number | string;
  recipient: string;
  due_date?: string | null;
  proof_id?: string | null;
}